Move glut gui into a project; and remove all references to glut in Sofa
