+++ applications/projects/SofaGuiGlut/Main.cpp
@@ -29,6 +29,11 @@ using std::string;
 #include <vector>
 using std::vector;
 
+#include "SimpleGUI.h"
+
+#ifdef SOFA_HAVE_BOOST
+#include "MultithreadGUI.h"
+#endif
 
 #include <sofa/helper/ArgumentParser.h>
 #include <SofaSimulationCommon/common.h>
@@ -62,7 +67,7 @@ using sofa::simulation::Node;
 using sofa::gui::GUIManager;
 
 #include <sofa/gui/Main.h>
-#include <sofa/gui/BatchGUI.h>  // For the default number of iterations
+#include <sofa/gui/BaseGUI.h>
 #include <sofa/helper/system/gl.h>
 #include <sofa/helper/system/atomic.h>
 
@@ -78,7 +83,6 @@ using sofa::simulation::tree::TreeSimulation;
 using sofa::simulation::graph::DAGSimulation;
 using sofa::helper::system::SetDirectory;
 using sofa::core::objectmodel::BaseNode ;
-using sofa::gui::BatchGUI;
 using sofa::gui::BaseGUI;
 
 #include <sofa/helper/logging/Messaging.h>
@@ -96,9 +100,6 @@ using  sofa::helper::logging::MainPerComponentLoggingMessageHandler ;
 #include <windows.h>
 #endif
 
-#include <sofa/gui/qt/GuiDataRepository.h>
-using sofa::gui::qt::GuiDataRepository ;
-
 using sofa::helper::system::DataRepository;
 using sofa::helper::system::PluginRepository;
 using sofa::helper::system::PluginManager;
@@ -115,39 +116,11 @@ using sofa::helper::logging::ExceptionMessageHandler;
 #define STRINGIFY(x) #x
 #define TOSTRING(x) STRINGIFY(x)
 
-
-void loadVerificationData(string& directory, string& filename, Node* node)
-{
-    msg_info("") << "loadVerificationData from " << directory << " and file " << filename ;
-
-    string refFile;
-
-    refFile += directory;
-    refFile += '/';
-    refFile += SetDirectory::GetFileName(filename.c_str());
-
-    msg_info("") << "loadVerificationData " << refFile ;
-
-    CompareStateCreator compareVisitor(ExecParams::defaultInstance());
-    compareVisitor.setCreateInMapping(true);
-    compareVisitor.setSceneName(refFile);
-    compareVisitor.execute(node);
-
-    ReadStateActivator v_read(ExecParams::defaultInstance(), true);
-    v_read.execute(node);
-}
-
-void addGUIParameters(ArgumentParser* argumentParser)
-{
-    GUIManager::RegisterParameters(argumentParser);
-}
-
 // ---------------------------------------------------------------------
 // ---
 // ---------------------------------------------------------------------
 int main(int argc, char** argv)
 {
-    GuiDataRepository.addFirstPath(Utils::getSofaPathTo("share/sofa/gui/runSofa/resources").c_str()) ;
     sofa::helper::BackTrace::autodump();
 
     ExecParams::defaultInstance()->setAspectID(0);
@@ -173,23 +146,13 @@ int main(int argc, char** argv)
     }
 #endif
 
-    sofa::gui::initMain();
-
     string fileName ;
-    bool        startAnim = false;
     bool        showHelp = false;
-    bool        printFactory = false;
+    bool        startAnim = false;
     bool        loadRecent = false;
-    bool        temporaryFile = false;
-    bool        testMode = false;
+    bool        printFactory = false;
     bool        noAutoloadPlugins = false;
-    unsigned int nbMSSASamples = 1;
-    bool computationTimeAtBegin = false;
-    unsigned int computationTimeSampling=0; ///< Frequency of display of the computation time statistics, in number of animation steps. 0 means never.
-    string    computationTimeOutputType="stdout";
-
-    string gui = "";
-    string verif = "";
+    bool        temporaryFile = false;
 
 #if defined(SOFA_HAVE_DAG)
     string simulationType = "dag";
@@ -207,40 +170,13 @@ int main(int argc, char** argv)
     string colorsStatus = "auto";
     string messageHandler = "auto";
     bool enableInteraction = false ;
+    string gui = "";
 
     string gui_help = "choose the UI (";
     gui_help += GUIManager::ListSupportedGUI('|');
     gui_help += ")";
 
     ArgumentParser* argParser = new ArgumentParser(argc, argv);
-    argParser->addArgument(po::value<bool>(&showHelp)->default_value(false)->implicit_value(true),                  "help,h", "Display this help message");
-    argParser->addArgument(po::value<bool>(&startAnim)->default_value(false)->implicit_value(true),                 "start,a", "start the animation loop");
-    argParser->addArgument(po::value<bool>(&computationTimeAtBegin)->default_value(false)->implicit_value(true),    "computationTimeAtBegin,b", "Output computation time statistics of the init (at the begin of the simulation)");
-    argParser->addArgument(po::value<unsigned int>(&computationTimeSampling)->default_value(0),                     "computationTimeSampling", "Frequency of display of the computation time statistics, in number of animation steps. 0 means never.");
-    argParser->addArgument(po::value<std::string>(&computationTimeOutputType)->default_value("stdout"),             "computationTimeOutputType,o", "Output type for the computation time statistics: either stdout, json or ljson");
-    argParser->addArgument(po::value<std::string>(&gui)->default_value(""),                                         "gui,g", gui_help.c_str());
-    argParser->addArgument(po::value<std::vector<std::string>>(&plugins),                                           "load,l", "load given plugins");
-    argParser->addArgument(po::value<bool>(&noAutoloadPlugins)->default_value(false)->implicit_value(true),         "noautoload", "disable plugins autoloading");
-
-    // example of an option using lambda function which ensure the value passed is > 0
-    argParser->addArgument(po::value<unsigned int>(&nbMSSASamples)->default_value(1)->notifier([](unsigned int value)
-    {
-        if (value < 1) {
-            std::cerr << "msaa sample cannot be lower than 1" << std::endl;
-            exit( EXIT_FAILURE );
-        }
-    }),                                                                                                             "msaa,m", "number of samples for MSAA (Multi Sampling Anti Aliasing ; value < 2 means disabled");
-
-    argParser->addArgument(po::value<bool>(&printFactory)->default_value(false)->implicit_value(true),              "factory,p", "print factory logs");
-    argParser->addArgument(po::value<bool>(&loadRecent)->default_value(false)->implicit_value(true),                "recent,r", "load most recently opened file");
-    argParser->addArgument(po::value<std::string>(&simulationType),                                                 "simu,s", "select the type of simulation (bgl, dag, tree)");
-    argParser->addArgument(po::value<bool>(&temporaryFile)->default_value(false)->implicit_value(true),             "tmp", "the loaded scene won't appear in history of opened files");
-    argParser->addArgument(po::value<bool>(&testMode)->default_value(false)->implicit_value(true),                  "test", "select test mode with xml output after N iteration");
-    argParser->addArgument(po::value<std::string>(&verif)->default_value(""), "verification,v",                     "load verification data for the scene");
-    argParser->addArgument(po::value<std::string>(&colorsStatus)->default_value("auto")->implicit_value("yes"),     "colors,c", "use colors on stdout and stderr (yes, no, auto)");
-    argParser->addArgument(po::value<std::string>(&messageHandler)->default_value("auto"), "formatting,f",          "select the message formatting to use (auto, clang, sofa, rich, test)");
-    argParser->addArgument(po::value<bool>(&enableInteraction)->default_value(false)->implicit_value(true),         "interactive,i", "enable interactive mode for the GUI which includes idle and mouse events (EXPERIMENTAL)");
-    argParser->addArgument(po::value<std::vector<std::string> >()->multitoken(), "argv",                            "forward extra args to the python interpreter");
 
 #ifdef SOFA_SMP
     argParser->addArgument(po::value<bool>(&disableStealing)->default_value(false)->implicit_value(true),           "disableStealing,w", "Disable Work Stealing")
@@ -248,14 +184,24 @@ int main(int argc, char** argv)
     argParser->addArgument(po::value<bool>(&affinity)->default_value(false)->implicit_value(true),                  "affinity", "Enable aFfinity base Work Stealing")
 #endif
 
-    addGUIParameters(argParser);
+
+    argParser->addArgument(po::value<bool>(&showHelp)->default_value(false)->implicit_value(true), "help,h", "Display this help message");
+    argParser->addArgument(po::value<bool>(&startAnim)->default_value(false)->implicit_value(true), "start,a", "start the animation loop");
+    argParser->addArgument(po::value<bool>(&printFactory)->default_value(false)->implicit_value(true), "factory,p", "print factory logs");
+    argParser->addArgument(po::value<bool>(&loadRecent)->default_value(false)->implicit_value(true), "recent,r", "load most recently opened file");
+    argParser->addArgument(po::value<bool>(&temporaryFile)->default_value(false)->implicit_value(true), "tmp", "the loaded scene won't appear in history of opened files");
+    argParser->addArgument(po::value<std::string>(&colorsStatus)->default_value("auto")->implicit_value("yes"), "colors,c", "use colors on stdout and stderr (yes, no, auto)");
+    argParser->addArgument(po::value<std::string>(&messageHandler)->default_value("auto"), "formatting,f", "select the message formatting to use (auto, clang, sofa, rich, test)");
+    argParser->addArgument(po::value<std::string>(&gui)->default_value(""), "gui,g", gui_help.c_str());
+    argParser->addArgument(po::value<bool>(&noAutoloadPlugins)->default_value(false)->implicit_value(true), "noautoload", "disable plugins autoloading");
+
     argParser->parse();
     files = argParser->getInputFileList();
 
-    if(showHelp)
+    if (showHelp)
     {
         argParser->showHelp();
-        exit( EXIT_SUCCESS );
+        exit(EXIT_SUCCESS);
     }
 
     // Note that initializations must be done after ArgumentParser that can exit the application (without cleanup)
@@ -269,6 +215,8 @@ int main(int argc, char** argv)
     sofa::component::initComponentGeneral();
     sofa::component::initComponentAdvanced();
     sofa::component::initComponentMisc();
+
+    glutInit(&argc, argv);
 
 #ifdef SOFA_HAVE_DAG
     if (simulationType == "tree")
@@ -325,7 +273,7 @@ int main(int argc, char** argv)
     for (unsigned int i=0; i<plugins.size(); i++)
         PluginManager::getInstance().loadPlugin(plugins[i]);
 
-    std::string configPluginPath = PluginRepository.getFirstPath() + "/" + TOSTRING(CONFIG_PLUGIN_FILENAME);
+    std::string configPluginPath = PluginRepository.getFirstPath() + "/" + TOSTRING(CONFIG_PLUGIN_FILENAME) ;
     std::string defaultConfigPluginPath = PluginRepository.getFirstPath() + "/" + TOSTRING(DEFAULT_CONFIG_PLUGIN_FILENAME);
 
     if (!noAutoloadPlugins)
@@ -350,7 +298,7 @@ int main(int argc, char** argv)
 
     PluginManager::getInstance().init();
 
-    if (int err = GUIManager::Init(argv[0],gui.c_str()))
+    if (int err = GUIManager::Init(argv[0], gui.c_str()))
         return err;
 
     if (fileName.empty())
@@ -379,24 +327,9 @@ int main(int argc, char** argv)
     if( !groot )
         groot = sofa::simulation::getSimulation()->createNewGraph("");
 
-    if (!verif.empty())
-    {
-        loadVerificationData(verif, fileName, groot.get());
-    }
-
-    if( computationTimeAtBegin )
-    {
-        sofa::helper::AdvancedTimer::setEnabled("Init", true);
-        sofa::helper::AdvancedTimer::setInterval("Init", 1);
-        sofa::helper::AdvancedTimer::setOutputType("Init", computationTimeOutputType);
-        sofa::helper::AdvancedTimer::begin("Init");
-    }
 
     sofa::simulation::getSimulation()->init(groot.get());
-    if( computationTimeAtBegin )
-    {
-        msg_info("") << sofa::helper::AdvancedTimer::end("Init", groot.get());
-    }
+
     GUIManager::SetScene(groot,fileName.c_str(), temporaryFile);
 
 
@@ -412,31 +345,14 @@ int main(int argc, char** argv)
         msg_info("") << "//////// END FACTORY ////////" ;
     }
 
-    if( computationTimeSampling>0 )
-    {
-        sofa::helper::AdvancedTimer::setEnabled("Animate", true);
-        sofa::helper::AdvancedTimer::setInterval("Animate", computationTimeSampling);
-        sofa::helper::AdvancedTimer::setOutputType("Animate", computationTimeOutputType);
-    }
-
     //=======================================
     // Run the main loop
     if (int err = GUIManager::MainLoop(groot,fileName.c_str()))
         return err;
     groot = dynamic_cast<Node*>( GUIManager::CurrentSimulation() );
 
-    if (testMode)
-    {
-        string xmlname = fileName.substr(0,fileName.length()-4)+"-scene.scn";
-        msg_info("") << "Exporting to XML " << xmlname ;
-        sofa::simulation::getSimulation()->exportXML(groot.get(), xmlname.c_str());
-    }
-
     if (groot!=NULL)
         sofa::simulation::getSimulation()->unload(groot);
-
-
-    GUIManager::closeGUI();
 
     sofa::simulation::common::cleanup();
     sofa::simulation::tree::cleanup();
